Add CartoDB map and slide/draw prototype
Proof of concept for the draw polygon/circle buttons
implemented on the home controller. Will need to add
to the museum controller as well.

TODO: It might be worth abstracting the draw trigger functions
but probably not since they'll only be used twice and are used
differently on both pages.