Search results to separate state
This commit moves the search results view to a separate
state so that we can use the browser back button to
navigate to it, removing the need for the awkward save
of the cookie state.

Making SearchController and DiscoverController is a bit
hacky, but seemed preferred to making directives that need
to pass a whole bunch of things around to preserve the map
and search states. These views are tightly coupled anyways.

- Geocoder searches init state transition to new search view
- Museum searches still transition directly to detail view
- Museum detail back button now mirrors browser back button
- Removed ngCookies dependency, no longer necessary
- Fixup map zoom/pan/marker on search results view
- Fix affix directive after nesting in ui-view