Move area analysis controls to map overlay
Constructs controls as a directive, interface using events.
List of available events are documented in
area-analysis-control-directive.js

While events are arguably not the best interface, this prevents
the trouble I ran into when I tried to nest angular directive
scope `'&'` calls to link the museum controller with the area controls
directive.