Load jQuery first
So Angular takes advantage of it everywhere.

Easiest way to do this was to specify jquery
as an explicit project dependency, and let
wiredep handle modification of the load order.

I don't see a quick, clean way digging into
wiredep settings to manually swap the order
without including jquery as a dep.